+++ NEWS.md
@@ -6,6 +6,14 @@ width: 128px;
 border-radius: 128px;
 " />
 
+## v1.12.5
+
+- Fixes
+    - Fix "Last Read" column in table view crashing.
+    - Ignore dotfiles.
+- Features
+    - Select by range in multi-select with shift+click.
+
 ## v1.12.4
 
 - Fixes